requirements.txt
pandas>=1.5.0
numpy>=1.23.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
pyyaml>=6.0
pytest>=7.0.0
